project-47-django-saas-ecommerce/requirements.txt
colorama==0.4.4
cryptography==41.0.1
distlib==0.3.6
docutils==0.16
jmespath==1.0.1
jsonschema==4.17.3
# mypy-extensions==1.0.0
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
resolvelib==1.0.1
urllib3==1.26.16
drf_yasg
python-decouple==3.1
unipath
django-cors-headers==4.2.0
celery
flower==1.2.0 
django_celery_results
django-celery-beat
djangorestframework
djangorestframework-simplejwt
redis
psycopg2-binary
django-redis==5.0.0
django-filter
pytest
pytest-django
graphene_django
requests
stripe
django-widget-tweaks
Pillow
django-debug-toolbar
django-data-browser
django-extensions
django-debug-toolbar-template-timings
django-countries
gunicorn